fixade teleporteringen samt lagget för nedre spelarna via udp

+++ Bowlerman/wall.c
@@ -179,19 +179,19 @@ void renderWalls(Game theGame)
     }
 }
 
-PUBLIC int getWallHeight(Wall w)
+PUBLIC float getWallHeight(Wall w)
 {
     return w.h;
 }
-PUBLIC int getWallWidth(Wall w)
+PUBLIC float getWallWidth(Wall w)
 {
     return w.w;
 }
-PUBLIC int getWallXPosition(Wall w)
+PUBLIC float getWallXPosition(Wall w)
 {
     return w.x;
 }
-PUBLIC int getWallYPosition(Wall w)
+PUBLIC float getWallYPosition(Wall w)
 {
     return w.y;
 }
